move alignment code to new module
